refactor waitForTemporalServerAndLog into TemporalUtils
